Possible fix
"screen" is already randomized in line 65, so there's no need to change it anymore here. In fact, it is incorrect to change it here, as it is a local variable and the NumberText was already set before this variable modification, meaning not only it won't be reflected in Start(), it also has no impact on NumberText.